Major Bug Fixes
The previous version did not display any results. This update has an alert-box pop-up to tell you if your guess is too high or too low. It will also display the final results of the game (whether you won or lost). If you lost, it will tell you the right answer.